requirements.txt
Office365_REST_Python_Client
pywin32
PyYAML
azure-identity